Update package.json
Hi guys! I was investigating on how to turn on sourcemaps support in node environment and I found one similar topic where it was pointed to use "-s" flag for source building to achieve that (https://medium.com/@Cuadraman/how-to-use-babel-for-production-5b95e7323c2f). Since your example mentioned in official babel docs, would be great to include it in the example.